add cylinder height and little fix
